+++ pkgs/test_core/lib/src/runner/engine.dart
@@ -597,7 +597,9 @@ class Engine {
   Future close() async {
     _closed = true;
     if (_closedBeforeDone != null) _closedBeforeDone = true;
-    await _suiteController.close();
+    if (_runCalled) {
+      await _suiteController.close();
+    }
     await _onSuiteAddedController.close();
 
     // Close the running tests first so that we're sure to wait for them to