+++ pkgs/test_core/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.6.20
 
+* Support `globalSetup` and `addGlobalTearDown` host runner execution and
+  caching across test suites.
 * Pass sanitizer environment defaults when running tests compiled as native CLI
   bundles.
 * Add support for `DART_TEST_REPORTER` environment variable.